Rename file and add missing returns annotation
